Rename the UploadDialog.js file to URLImportDialog.js
